Add files via upload
Performed predictive modelling on dataset of Manipal Hospital so that the customer experience at Manipal Hospital can be increased by correcting the areas of concern in real time. The dataset had 51 variables which include customer’s feedback on some specific departments and also the answer to NPS question (How likely are they to recommend Hospital to friends and family). Pre-processing is done by transforming imbalance in the data via deploying techniques like SMOTE, Up-sampling and Down-sampling. After that Applied different Ensemble models like Random Forest and Adaptive boosting to predict class of the patient in real time which will help us to take detractors(who is not supporter) to desire class by implementing measures for the problems which will eventually improve NPS score.